+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+
+## [4.16.9][] - 2017-05-07
 ### Changed
 - Extracts react-scroll-paginator
 
@@ -701,7 +703,8 @@ At a later date a embed generator page will be made to help with the creation of
 ### Added
 - This CHANGELOG file to allow more insight to the changes made throughout the development of gw2armory.com
 
-[Unreleased]: https://github.com/madou/armory-react/compare/v4.16.8...HEAD
+[Unreleased]: https://github.com/madou/armory-react/compare/v4.16.9...HEAD
+[4.16.9]: https://github.com/madou/armory-react/compare/v4.16.8...v4.16.9
 [4.16.8]: https://github.com/madou/armory-react/compare/v4.16.7...v4.16.8
 [4.16.7]: https://github.com/madou/armory-react/compare/v4.16.6...v4.16.7
 [4.16.6]: https://github.com/madou/armory-react/compare/v4.16.5...v4.16.6